x1plusd.actions: move delay, gcode, file back into core
