+++ images/cfw/opt/x1plus/lib/python/x1plus/services/x1plusd/actions.py
@@ -95,10 +95,50 @@ def register_action(name, handler = None):
     def decorator(handler):
         assert name not in _registered_actions
         _registered_actions[name] = handler
-        logger.info(f"Registered Actions from module: {name}")
+        logger.info(f"registered Action handler \"{name}\"")
         return handler
 
     if handler is None:
         return decorator
     else:
         decorator(handler)
+
+###
+
+@register_action("delay")
+async def _action_delay(handler, subconfig):
+    logger.debug(f"delay action: {subconfig}")
+    if type(subconfig) != int and type(subconfig) != float:
+        raise TypeError(f"delay parameter {subconfig} was not numberish")
+    await asyncio.sleep(subconfig)
+
+@register_action("gcode")
+async def _action_gcode(handler, subconfig):
+    logger.debug(f"gcode action: {subconfig}")
+    if type(subconfig) != str:
+        raise TypeError(f"gcode parameter {subconfig} was not str")
+    await handler.daemon.mqtt.publish_request({ "print": { "command": "gcode_line", "sequence_id": "0", "param": subconfig } })
+
+@register_action("file")
+async def _action_file(handler, subconfig):
+    if type(subconfig) != str:
+        raise TypeError(f"parameter {subconfig} to 'file' action was not a string")
+    if subconfig[0] != "/":
+        subconfig = f"/sdcard/{subconfig}"
+    with open(subconfig) as f:
+        contents_raw = f.read()
+    
+    contents = None
+    if contents is None:
+        try:
+            contents = json.loads(contents_raw)
+        except:
+            pass
+    if contents is None:
+        try:
+            contents = yaml.safe_load(contents_raw)
+        except:
+            pass
+    if contents is None:
+        raise ValueError(f"file {subconfig} seemed to be neither a yaml file nor a json file")
+    return await handler.execute_step(contents)
